# api/dts.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "dts.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/dependencies": {
      "get": {
        "operationId": "getDependencies",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/apiResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "organization",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "module",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "DOWNSTREAM",
              "UPSTREAM"
            ],
            "default": "DOWNSTREAM"
          }
        ],
        "tags": [
          "DependencyTracker"
        ]
      }
    },
    "/v1/dependencies/managed": {
      "get": {
        "operationId": "getManaged",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetManagedResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DependencyTracker"
        ]
      }
    },
    "/v1/dependencies/track": {
      "post": {
        "operationId": "put",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiPutResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiPutRequest"
            }
          }
        ],
        "tags": [
          "DependencyTracker"
        ]
      }
    },
    "/v1/languages": {
      "get": {
        "operationId": "listLanguages",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiListLanguagesResponse"
            }
          }
        },
        "tags": [
          "DependencyTracker"
        ]
      }
    },
    "/v1/modules": {
      "get": {
        "operationId": "listModules",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiListModulesResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "organization",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DependencyTracker"
        ]
      }
    },
    "/v1/organizations": {
      "get": {
        "operationId": "listOrganizations",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiListOrganizationsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DependencyTracker"
        ]
      }
    },
    "/v1/sources": {
      "get": {
        "operationId": "getSources",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/apiGetSourcesResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "organization",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "module",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "DependencyTracker"
        ]
      }
    },
    "/v1/topology": {
      "get": {
        "operationId": "getTopology",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/apiResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "organization",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "module",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "DOWNSTREAM",
              "UPSTREAM"
            ],
            "default": "DOWNSTREAM"
          }
        ],
        "tags": [
          "DependencyTracker"
        ]
      }
    },
    "/v1/topology/tiered": {
      "get": {
        "operationId": "getTopologyTiered",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/apiTieredResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "organization",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "module",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "DOWNSTREAM",
              "UPSTREAM"
            ],
            "default": "DOWNSTREAM"
          }
        ],
        "tags": [
          "DependencyTracker"
        ]
      }
    }
  },
  "definitions": {
    "apiDependency": {
      "type": "object",
      "properties": {
        "organization": {
          "type": "string"
        },
        "module": {
          "type": "string"
        },
        "versionConstraint": {
          "type": "string"
        },
        "scopes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiDependencyId": {
      "type": "object",
      "properties": {
        "language": {
          "type": "string"
        },
        "organization": {
          "type": "string"
        },
        "module": {
          "type": "string"
        }
      }
    },
    "apiDependencyManagementFile": {
      "type": "object",
      "properties": {
        "language": {
          "type": "string"
        },
        "system": {
          "type": "string"
        },
        "organization": {
          "type": "string"
        },
        "module": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "dependencies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiDependency"
          }
        }
      }
    },
    "apiDirection": {
      "type": "string",
      "enum": [
        "DOWNSTREAM",
        "UPSTREAM"
      ],
      "default": "DOWNSTREAM"
    },
    "apiGetManagedResponse": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        },
        "managed": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiDependencyId"
          }
        }
      }
    },
    "apiGetSourcesResponse": {
      "type": "object",
      "properties": {
        "source": {
          "$ref": "#/definitions/apiSourceInformation"
        }
      }
    },
    "apiListLanguagesResponse": {
      "type": "object",
      "properties": {
        "languages": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiListModulesResponse": {
      "type": "object",
      "properties": {
        "language": {
          "type": "string"
        },
        "organization": {
          "type": "string"
        },
        "modules": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiListOrganizationsResponse": {
      "type": "object",
      "properties": {
        "language": {
          "type": "string"
        },
        "organizations": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apiPutRequest": {
      "type": "object",
      "properties": {
        "sourceInformation": {
          "$ref": "#/definitions/apiSourceInformation"
        },
        "managementFiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiDependencyManagementFile"
          }
        }
      }
    },
    "apiPutResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "apiResponse": {
      "type": "object",
      "properties": {
        "dependency": {
          "$ref": "#/definitions/apiDependencyId"
        }
      }
    },
    "apiSourceInformation": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        }
      }
    },
    "apiTieredResponse": {
      "type": "object",
      "properties": {
        "tier": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiDependencyId"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeStreamError": {
      "type": "object",
      "properties": {
        "grpc_code": {
          "type": "integer",
          "format": "int32"
        },
        "http_code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "http_status": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "x-stream-definitions": {
    "apiGetSourcesResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/apiGetSourcesResponse"
        },
        "error": {
          "$ref": "#/definitions/runtimeStreamError"
        }
      },
      "title": "Stream result of apiGetSourcesResponse"
    },
    "apiResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/apiResponse"
        },
        "error": {
          "$ref": "#/definitions/runtimeStreamError"
        }
      },
      "title": "Stream result of apiResponse"
    },
    "apiTieredResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/apiTieredResponse"
        },
        "error": {
          "$ref": "#/definitions/runtimeStreamError"
        }
      },
      "title": "Stream result of apiTieredResponse"
    }
  }
}